requirements.txt
Django~=4.2.16
drf-spectacular~=0.27.2
djangorestframework~=3.15.2
django-environ~=0.11.2
django-cors-headers~=4.5.0
pytils~=0.4.1
django-filter~=24.3